disallow functions in calls to math-read-expr
This is achieved through a new dynamic variable
symtex--calc-allow-functions, which can be let-bound to nil to disable
functions.  The story here is that I often write stuff in latex like
"x (y+1)" which I want to interpret as a product involving algebraic
quantities.  By default, calc parses this as the function x evaluated
at the argument y+1.  I never use symtex for functions, so it makes
sense to tweak the parsing as I've done here.